refactor for new view layout
