+++ src/app/Utils/objects/breakageInfo.ts
