+++ src/app/Utils/objects/docLink.ts
