+++ src/app/shared/breakage-card/breakage-card.component.html
