+++ src/app/shared/dialog/image-modal.component.ts
