+++ src/app/view-report/report-usage/last-uses/last-uses.component.ts
@@ -1,9 +1,9 @@
 import { Component, OnInit, Input } from '@angular/core';
 
-import { BoatUsageService } from '../boat-usage.service'
-import { UsageInfo } from '../objects/usageInfo'
+import { BoatUsageService } from '../../../boat-usage.service'
+import { UsageInfo } from '../../../Utils/objects/usageInfo'
 
-import { BoatNameConversionHelper } from '../Utils/nameConversion'
+import { BoatNameConversionHelper } from '../../../Utils/nameConversion'
 
 @Component({
   selector: 'last-uses',
