+++ src/app/view-stats/stats.component.html
