+++ src/app/view-stats/stats.component.spec.ts
