+++ src/app/view-stats/stats.component.ts
