+++ src/app/view-view/view-fixed/view-fixed.component.spec.ts
