+++ src/app/view-view/view.component.ts
